requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
xgboost==2.0.0
supabase==2.0.1
apscheduler==3.10.4
gunicorn==21.2.0